\name{print.summary.pop}
\alias{print.summary.pop}
\title{Print population summary}
\usage{
  print.summary.pop(object)
}
\arguments{
  \item{object}{the summary to print}
}
\value{
  returns the object in the call invisibly
}
\description{
  Print the summary of a population
}
\details{
  just pretty formatting
}
\author{
  Christoph Waldhauser
}

