\name{summary.pop}
\alias{summary.pop}
\title{Summarize a population}
\usage{
  \method{summary}{pop} (object, ...)
}
\arguments{
  \item{object}{population}

  \item{...}{currently ignored}
}
\value{
  a list with summary information
}
\description{
  Produces a summary of a population
}
\details{
  Fields returned are type, size, amount of Hs and Zs and
  the mean infection rate.
}
\author{
  Christoph Waldhauser
}

